Moved the drum for playing in its own app state
